symbol by pool pair

+++ contracts/interfaces/limit/ILimitPoolFactory.sol
@@ -16,7 +16,7 @@ abstract contract ILimitPoolFactory is LimitPoolFactoryStorage, PoolsharkStructs
         address tokenIn,
         address tokenOut,
         uint16  swapFee,
-        uint8   poolTypeId
+        uint16  poolTypeId
     ) external view virtual returns (
         address pool,
         address poolToken
